Split inspection module responsibilities
